Add the marker chart search filtering
This PR has a bit larger of a code change than just the filtering, as
the selectors needed to be reworked a little bit to properly handle
filtering by searches. In addition, the routing for network markers in
the selectors was breaking memoization, when alternating calls for
tracing markers between the various panels.