Add a restricted height and width strategy for tooltips
This option is set up in the props, as it will be enabled for tooltips,
and disabled for the sidebar. Restricting the width of the tooltip helps
it to be more stable in its layout across different markers. This
creates a nicer experience for the user.

The stacks were using a fading out strategy, but I felt it was better to
just provide an ellipsis "..." in order to show that the stack had been
truncated. In addition, I moved around the stack to be more in line with
the other pieces of the tooltip.