(Do not review, PR #3050) Make profile processing a bit more type safe
This patch reworks our processing pipeline a bit, to not assume too
much.

It uses more mixed types rather than the Object type. It also embeds the
if checks for profile types into the actual processing function.
Previously we weren't benefiting from the type refinement that is only
available inline here.

This should have the side effect of having better runtime errors for
things like e.meta is not defined.

Resolves #1947